+++ Editor/EditorUtils/VXEditorUtil.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Reflection;
 using UnityEngine;
 using UnityEditor;
+using UnityEditorInternal;
 
 namespace Voxell.Inspector
 {