+++ CHANGELOG.md
@@ -4,16 +4,21 @@
 
 - Added inverse lerp functions to support `float`, `float2`, and `float4` operations.
 - Added default fields to `VXEditorUtil`.
-- `FoldableReorderableList`:
+- `CreateReorderableList` (formerly known as `FoldableReorderableList`):
   - Supports serialized structs/classes.
   - Element height now depends on serialized property height.
   - Added "empty message" to display when the list is empty.
+- Created `GroupableList` class for storing a group index for each reorderable list.
+- Created `EditorListConfig` struct to store `ReorderableList` related data.
 
 ### Changes
 
 - Removed `AbstractVXScriptableEditor`.
 - `serializedObject` now updates every frame in `OnInspectorGUI` (in `AbstractVXEditor`) to refresh `serializedObject` representation (especially in scriptable objects).
 - Removed unused functions from `VXEditorUtil`.
+- Renamed `FoldableReorderableList` to `CreateReorderableList`.
+- Added new `namespace` -> `Voxell.Inspector.List` for list specific classes/structs.
+- Moved `CreateReorderableList` to `EditorListUtil` file.
 
 ## [1.4.1]